# docs/swagger.yaml
basePath: /api
definitions:
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel:
    properties:
      code: {}
      message:
        example: error
        type: string
      status:
        example: error
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel:
    properties:
      errors: {}
      message:
        example: validation error
        type: string
      status:
        example: error
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel:
    properties:
      contents: {}
      kind:
        example: collection
        type: string
      total_rows:
        example: 100
        type: integer
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.ConfigVatRevenue:
    properties:
      endTime:
        type: string
      percentage:
        type: number
      startTime:
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.OrderType:
    properties:
      orderTypeCode:
        type: string
      orderTypeName:
        type: string
      transactionTypes:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionType'
        type: array
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.OrderTypeOut:
    properties:
      kind:
        type: string
      orderTypeCode:
        type: string
      orderTypeName:
        type: string
      transactionTypes:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionTypeOut'
        type: array
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionType:
    properties:
      transactionTypeCode:
        type: string
      transactionTypeName:
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionTypeOut:
    properties:
      kind:
        type: string
      transactionTypeCode:
        type: string
      transactionTypeName:
        type: string
    type: object
...
host: localhost:9567
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a go fp transaction api docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GO FP TRANSACTION API DUCUMENTATION
  version: "1.0"
paths:
...
  /v1/order-types:
    get:
      consumes:
      - application/json
      description: Get all data order type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all data order type
      tags:
      - OrderType
    post:
      consumes:
      - application/json
      description: Create data order type
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.OrderType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Create data order type
      tags:
      - OrderType
  /v1/order-types/:orderTypeCode:
    get:
      consumes:
      - application/json
      description: Get order type detail
      parameters:
      - description: order type code
        in: path
        name: orderTypeCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.OrderTypeOut'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get order type detail
      tags:
      - OrderType
...
  /v1/transaction-types:
    get:
      consumes:
      - application/json
      description: Get all data transaction type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all data transaction type
      tags:
      - TransactionType
  /v1/transaction-types/:transactionTypeCode:
    get:
      consumes:
      - application/json
      description: Get transaction type detail
      parameters:
      - description: transaction type code
        in: path
        name: transactionTypeCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionTypeOut'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get transaction type detail
      tags:
      - TransactionType
...
  /v1/vat-configs:
    get:
      consumes:
      - application/json
      description: Get all data vat config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all data vat config
      tags:
      - VATConfig
    patch:
      consumes:
      - application/json
      description: Update data vat config
      parameters:
      - description: body
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.ConfigVatRevenue'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Update data vat config
      tags:
      - VATConfig
...
schemes:
- http
swagger: "2.0"